+++ poulpy-cpu-avx/src/ntt120/tests.rs
@@ -1,7 +1,9 @@
 use poulpy_hal::{backend_test_suite, cross_backend_test_suite};
 use poulpy_hal::{
     layouts::Module,
-    test_suite::convolution::{test_convolution, test_convolution_by_const, test_convolution_pairwise},
+    test_suite::convolution::{
+        test_convolution, test_convolution_accumulate, test_convolution_by_const, test_convolution_pairwise,
+    },
 };
 
 use crate::NTT120Avx;
@@ -201,6 +203,7 @@ fn test_convolution_direct() {
     test_convolution(&module, 50);
     test_convolution_by_const(&module, 50);
     test_convolution_pairwise(&module, 50);
+    test_convolution_accumulate(&module, 50);
 }
 
 #[test]